Move the pull request fetching up one level
